requirements.txt
streamlit==1.38.0
numpy==1.26.3
pandas==2.2.0
pydeck==0.9.1
altair==5.4.1
vega_datasets==0.9.0
plotly==5.24.1
scipy==1.12.0
pillow==10.4.0

